+++ Runtime/Core/Platform/IPlatformImpl.cs
@@ -112,6 +112,27 @@ namespace Unity.AppUI.Core
         Color GetSystemColor(SystemColorType colorType);
 
         /// <summary>
+        /// Whether the platform has data of the given type in the pasteboard.
+        /// </summary>
+        /// <param name="type"> The type of data to check for. </param>
+        /// <returns> Whether the platform has data of the given type in the pasteboard. </returns>
+        bool HasPasteboardData(PasteboardType type);
+
+        /// <summary>
+        /// Get the data of the given type to the pasteboard.
+        /// </summary>
+        /// <param name="type"> The type of data to get. </param>
+        /// <returns> The data of the given type from the pasteboard. </returns>
+        byte[] GetPasteboardData(PasteboardType type);
+
+        /// <summary>
+        /// Set the data of the given type to the pasteboard.
+        /// </summary>
+        /// <param name="type"> The type of data to set. </param>
+        /// <param name="data"> The data to set. </param>
+        void SetPasteboardData(PasteboardType type, byte[] data);
+
+        /// <summary>
         /// Event to update the native integration.
         /// </summary>
         /// <remarks>